backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-cloud-aiplatform
google-generativeai
googlemaps
firebase-admin
google-cloud-bigquery
stripe
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
celery
sqlalchemy
alembic
psycopg2-binary
aiofiles
pandas
numpy
